Fixed blog editing

+++ app/controllers/blog_controller.php
@@ -9,13 +9,15 @@ class BlogController extends BaseController{
     
     public static function show($id) {
         $blog = Blog::find($id);
-        $posts = Post::allInBlog($id);       
-   	    View::make('blog/show.html', array('blog' => $blog, 'posts' => $posts));
+        $posts = Post::allInBlog($id);
+   	    View::make('blog/show.html', array('blog' => $blog, 'posts' => $posts,
+           'can_destroy' => Blog::canDestroy($blog->blogId)));
     }
     
     public static function edit($id) {
         $blog = Blog::find($id);
-   	    View::make('blog/edit.html', array('blog' => $blog));
+   	    View::make('blog/edit.html', array('blog' => $blog, 'can_destroy' => Blog::canDestroy($blog->blogId),
+           'can_edit' => Blog::canEdit($blog->blogId)));
     }
     
     public static function create() {
@@ -29,18 +31,28 @@ class BlogController extends BaseController{
                 'description' => trim($_POST['description'])
         ));
         
+        if (!Blog::canEdit($blog->blogId)) {
+            Redirect::to('/blog/' . $blog->blogId, array('errors' => array('Ei ole vaadittavia oikeuksia.')));
+            return;
+        }
+        
         $errors = $blog->errors();
 
-        if (!empty($errors)){
+        if (count($errors) == 0) { // !empty($errors) didn't work
             $blog->update();
             Redirect::to('/blog/' . $blog->blogId, array('message' => 'Blogia on muokattu!'));
         } else {
-            View::make('blog/edit.html', array('errors' => $errors));
+            Redirect::to('/blog/' . $blog->blogId . '/edit', array('errors' => $errors));
         }
     }
 
     
     public static function destroy($id) {
+        if (!Blog::canDestroy($id)) {
+            Redirect::to('/blog/' . $blog->blogId);
+            return;
+        }
+        
         Blog::destroy($id);
         Redirect::to('/blog');
     }
